LayoutTests/media/modern-media-controls/tracks-support/audio-text-tracks-not-shown.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<meta name="viewport" content="width=device-width">
<script src="../resources/media-controls-utils.js"></script>
<script src="../../../resources/ui-helper.js"></script>
<script src="../../../resources/js-test.js"></script>
<body>
<audio style="position: absolute; left: 0; top: 0; width: 600px;" controls>
    <source src="../../content/counting.mp4">
    <track src="../../content/lorem-ipsum.vtt" kind="subtitles" srclang="en" label="English" default>
</audio>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Check that subtitle tracks are not offered for audio elements.");

const media = document.querySelector("audio");
const shadowRoot = window.internals.shadowRoot(media);

window.internals.setMediaControlsMaximumRightContainerButtonCountOverride(media, 100);

let contextmenu = null;

media.addEventListener("canplay", async () => {
    await shouldBecomeDifferent("shadowRoot.querySelector('button.overflow')", "null");
    await shouldBecomeDifferent("shadowRoot.querySelector('button.overflow').getBoundingClientRect().width", "0");

    debug("Tapping overflow button...");
    await pressOnElement(shadowRoot.querySelector("button.overflow"));
    contextmenu = await getTracksContextMenu();
    if (contextmenu.length === 1 && !contextmenu[0].title && contextmenu[0].children?.length)
        contextmenu = contextmenu[0].children;
    shouldBeFalse("contextmenu.some((item) => item.title === 'Subtitles')");

    debug("Dismissing contextmenu...");
    await new Promise((resolve, reject) => {
        testRunner.runUIScript(`
            uiController.dismissMenu();
            uiController.uiScriptComplete();
        `, resolve);
    });
    await UIHelper.waitForContextMenuToHide();

    await shouldBecomeDifferent("shadowRoot.querySelector('button.tracks')", "null");
    await shouldBecomeEqual("shadowRoot.querySelector('button.tracks').getBoundingClientRect().width", "0");

    media.remove();
    finishJSTest();
});

</script>
</body>
