LayoutTests/media/modern-media-controls/tracks-support/audio-text-tracks-not-shown-expected.txt

Check that subtitle tracks are not offered for audio elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS shadowRoot.querySelector('button.overflow') became different from null
PASS shadowRoot.querySelector('button.overflow').getBoundingClientRect().width became different from 0
Tapping overflow button...
PASS contextmenu.some((item) => item.title === 'Subtitles') is false
Dismissing contextmenu...
PASS shadowRoot.querySelector('button.tracks') became different from null
PASS shadowRoot.querySelector('button.tracks').getBoundingClientRect().width became 0
PASS successfullyParsed is true

TEST COMPLETE
